Avoid duplicated Actions runs and deploy only from main (#580)
Update run conditions for GitHub Actions. Run GitHub Actions on every
Pull Request and only on pushes to `main` branch. This avoids having
duplicated runs in every PR: one because it's a PR and one because it's
a push to a branch. Add condition to deploy GitHub Pages only if the
built came from a push to `main`.