Use geoana 0.2.1 for building the website (#579)
Use an older version of geoana to build the website to avoid errors
coming from incompatibilities with the latest version. Install
`geoana==0.2.1` in GitHub Actions to run tests and to build the website.
Just pinning it in the `requirements.txt` file doesn't work because of
missing build dependencies for that old version. So the build
requirements (listed in a new `requirements-build-geoana.txt` file) are
installed along with other needed packages (e.g. Sphinx), and
`geoana=0.2.1` gets installed with `pip` but disabling the build
isolation (passing `--no-build-isolation`).

This fix is just a patch, we should make the repo compatible with latest
geoana at some point. By then, we should revert these changes.